backend/requirements.txt
streamlit>=1.28.0
crewai>=0.1.0
fastapi>=0.104.0
uvicorn>=0.24.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
pydantic>=2.5.0
python-dotenv>=1.0.0
groq>=0.4.0
requests>=2.31.0
python-multipart>=0.0.6
PyPDF2>=3.0.0
python-docx>=1.1.0


